intel/isl: Use uint64_t for computed byte offsets
This is mostly a bit of future-proofing.  We never end up with offsets
that don't fit in 32 bits today because, thanks to driver limitations
caused by relocations, we don't allocate buffers bigger than 2GB today.
However, if we ever did, it's possible to create a surface on modern
platforms that consumes more than 4GB and we would end up with wrapping
in our offset calculations.

Acked-by: Ivan Briano <ivan.briano@intel.com>
Part-of: <https://gitlab.freedesktop.org/mesa/mesa/-/merge_requests/11765>

+++ src/intel/blorp/blorp_genX_exec.h
@@ -1657,7 +1657,7 @@ blorp_emit_depth_stencil_config(struct blorp_batch *batch,
           * anyway by manually offsetting to the specified miplevel.
           */
          assert(info.hiz_surf->dim_layout == ISL_DIM_LAYOUT_GFX6_STENCIL_HIZ);
-         uint32_t offset_B;
+         uint64_t offset_B;
          isl_surf_get_image_offset_B_tile_sa(info.hiz_surf,
                                              info.view->base_level, 0, 0,
                                              &offset_B, NULL, NULL);
@@ -1683,7 +1683,7 @@ blorp_emit_depth_stencil_config(struct blorp_batch *batch,
        * anyway by manually offsetting to the specified miplevel.
        */
       assert(info.stencil_surf->dim_layout == ISL_DIM_LAYOUT_GFX6_STENCIL_HIZ);
-      uint32_t offset_B;
+      uint64_t offset_B;
       isl_surf_get_image_offset_B_tile_sa(info.stencil_surf,
                                           info.view->base_level, 0, 0,
                                           &offset_B, NULL, NULL);
